moved feedback model into feedback module
