+++ src/app/feedback/views.py
@@ -1,23 +1,19 @@
 from django.conf import settings
-
 from rest_framework.views import APIView
+from .models import Feedback
 
-from . import local
-
-
-class GetFeedback(APIView):
-    """Upload file on the S3 storage."""
+class FeedbackThumbsUp(APIView):
     def post(self, request, **kwargs):
-        """Accept file uploads to AWS and local filesystem."""
-        if settings.S3_STORAGE:
-            return aws.post(request)
-        else:
-            return local.post(request)
-
+        return self.create(request, *args, **kwargs)
 
     def get(self, request, **kwargs):
-        """List doc files from S3 or local file system."""
-        if settings.S3_STORAGE:
-            return aws.get(request)
-        else:
-            return local.get(request)
+        return self.list(request, *args, **kwargs)
+
+
+class FeedbackThumbsDown(APIView):
+    def post(self, request, **kwargs):
+        return self.create(request, *args, **kwargs)
+
+    def get(self, request, **kwargs):
+        return self.list(request, *args, **kwargs)
+
